fix: PostgreSQL ROUND function compatibility

+++ src/routes/admin.js
@@ -233,11 +233,11 @@ router.get('/filtering-test/:accountId', async (req, res) => {
       )
       SELECT 
         *,
-        CAST(ROUND((conversation_emails::float / NULLIF(total_emails, 0)) * 100, 2) AS NUMERIC(5,2)) as conversation_percentage,
-        CAST(ROUND((valid_for_processing::float / NULLIF(pending_emails, 0)) * 100, 2) AS NUMERIC(5,2)) as valid_processing_percentage,
+        CAST(ROUND((conversation_emails::NUMERIC / NULLIF(total_emails, 0)) * 100, 2) AS NUMERIC(5,2)) as conversation_percentage,
+        CAST(ROUND((valid_for_processing::NUMERIC / NULLIF(pending_emails, 0)) * 100, 2) AS NUMERIC(5,2)) as valid_processing_percentage,
         (pending_emails - valid_for_processing) as emails_filtered_out,
-        CASE 
-          WHEN pending_emails > 0 THEN ROUND(((pending_emails - valid_for_processing)::float / pending_emails) * 100, 2)
+        CASE
+          WHEN pending_emails > 0 THEN ROUND(((pending_emails - valid_for_processing)::NUMERIC / pending_emails) * 100, 2)
           ELSE 0
         END as spam_reduction_percentage
       FROM email_analysis
